Result from all the implemented methods

+++ main.py
@@ -7,6 +7,7 @@ from logging import INFO
 import os
 import sys
 import yaml
+import json
 import glob
 import pickle
 import torch
@@ -103,13 +104,15 @@ def mainFedOAP(args,cfg):
     temporaryWeightsPaths = glob.glob('temporaryWeights/*.pth')
     for path in temporaryWeightsPaths:
       os.remove(path)
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
     exit()
   finally:
     temporaryWeightsPaths = glob.glob('temporaryWeights/*.pth')
     for path in temporaryWeightsPaths:
       os.remove(path)
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
 
   try:
     log(INFO, "Going into Finetuning")
@@ -123,7 +126,8 @@ def mainFedOAP(args,cfg):
       epochs=cfg['finetuning_epochs'],
       val_per_epoch=cfg['val_per_epoch'],
       in_channels=cfg['input_channels'],
-      num_classes=cfg['num_classes']
+      num_classes=cfg['num_classes'],
+      run=args.run
     )
 
     trainer.train()
@@ -133,8 +137,10 @@ def mainFedOAP(args,cfg):
     traceback.print_exc()
     for idx in range(cfg['num_clients']):
       finetunedName = f'fedOAPfinetuned{idx}.pth'
-      os.remove(os.path.join(cfg['output_dir'],finetunedName))
-    os.remove(os.path.join(cfg['output_dir'],'results.txt'))
+      if os.path.exists(os.path.join(cfg['output_dir'],finetunedName)):
+        os.remove(os.path.join(cfg['output_dir'],finetunedName))
+    if os.path.exists(os.path.join(cfg['output_dir'],f'results{args.run}.txt')):
+      os.remove(os.path.join(cfg['output_dir'],f'results{args.run}.txt'))
 
   del trainer
 
@@ -154,6 +160,11 @@ def mainFedDP(args,cfg):
   # Load Datasets
   datasets = load_datasets(cfg['dataset_dirs'], cfg['image_size'])
   log(INFO, "Datasets loaded. Number of datasets: %s", len(datasets))
+  log(INFO, "Number of Clients: %s", len(datasets))
+
+  for ix in range(len(datasets)):
+    log(INFO,f'Number of samples for client {ix} : {datasets[ix].__len__()}')
+  
   train_dataloaders, val_dataloaders, test_dataloaders = prepare_datasets(
     datasets=datasets, 
     batch_size=cfg['batch_size'], 
@@ -210,13 +221,15 @@ def mainFedDP(args,cfg):
     temporaryWeightsPaths = glob.glob('temporaryWeights/*.pth')
     for path in temporaryWeightsPaths:
       os.remove(path)
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
     exit()
   finally:
     queryWeightsPaths = glob.glob('temporaryWeights/*.pth')
     for weight_path in queryWeightsPaths:
       os.remove(weight_path)
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
 
 
   try:
@@ -231,7 +244,8 @@ def mainFedDP(args,cfg):
       epochs=cfg['igc_epochs'],
       val_per_epoch=cfg['val_per_epoch'],
       in_channels=cfg['input_channels'],
-      num_classes=cfg['num_classes']
+      num_classes=cfg['num_classes'],
+      run=args.run
     )
 
     trainer.train()
@@ -241,8 +255,10 @@ def mainFedDP(args,cfg):
     traceback.print_exc()
     for idx in range(cfg['num_clients']):
       finetunedName = f'fedDPfinetuned{idx}.pth'
-      os.remove(os.path.join(cfg['output_dir'],finetunedName))
-    os.remove(os.path.join(cfg['output_dir'],'results.txt'))
+      if os.path.exists(os.path.join(cfg['output_dir'],finetunedName)):
+        os.remove(os.path.join(cfg['output_dir'],finetunedName))
+    if os.path.exists(os.path.join(cfg['output_dir'],f'results{args.run}.txt')):
+      os.remove(os.path.join(cfg['output_dir'],f'results{args.run}.txt'))
 
   del trainer
 
@@ -346,7 +362,8 @@ def mainFedREP(args,cfg):
       epochs=cfg['finetune_epochs'],
       val_per_epoch=cfg['val_per_epoch'],
       in_channels=cfg['input_channels'],
-      num_classes=cfg['num_classes']
+      num_classes=cfg['num_classes'],
+      run=args.run
     )
 
     trainer.train()
@@ -356,9 +373,10 @@ def mainFedREP(args,cfg):
     traceback.print_exc()
     for idx in range(cfg['num_clients']):
       finetunedName = f'fedREPfinetuned{idx}.pth'
-      os.remove(os.path.join(cfg['output_dir'],finetunedName))
-    if os.path.exists(os.path.join(cfg['output_dir'],'results.txt')):
-      os.remove(os.path.join(cfg['output_dir'],'results.txt'))
+      if os.path.exists(os.path.join(cfg['output_dir'],finetunedName)):
+        os.remove(os.path.join(cfg['output_dir'],finetunedName))
+    if os.path.exists(os.path.join(cfg['output_dir'],f'results{args.run}.txt')):
+      os.remove(os.path.join(cfg['output_dir'],f'results{args.run}.txt'))
 
   del trainer
 
@@ -456,8 +474,12 @@ def mainFedPER(args,cfg):
       in_channels=cfg['input_channels'],
       num_classes=cfg['num_classes']
     )
-    test_model.load_state_dict(torch.load(os.path.join(cfg['output_dir'],'fedPERserver.pth')))
-    test_model.head.load_state_dict(torch.load(os.path.join(cfg['output_dir'],f'fedPERhead{idx}.pth')))
+    test_model.load_state_dict(
+      torch.load(os.path.join(cfg['output_dir'],'fedPERserver.pth'))
+    )
+    test_model.head.load_state_dict(
+      torch.load(os.path.join(cfg['output_dir'],f'fedPERhead{idx}.pth'))
+    )
     test_model.eval()
     test_model.to(device)
     dice_scores = []
@@ -471,10 +493,15 @@ def mainFedPER(args,cfg):
         dice_scores.append(dice)
     results.append(torch.tensor(dice_scores).mean().item())
     
-  with open(os.path.join(cfg['output_dir'],'results.txt'),'w') as f:
-    for idx in range(cfg['num_clients']):
-      if idx > 0 : f.write(f'\nclient {idx} has dice score :{results[idx]}')
-      else : f.write(f'client {idx} has dice score {results[idx]}')
+  result_dict = {
+    '0':-1.0,
+    '1':-1.0,
+    '2':-1.0
+  }
+  for idx in range(cfg['num_clients']):
+      result_dict[str(idx)] = results[idx]
+  with open(os.path.join(cfg['output_dir'],f'results{args.run}.json'), "w") as f:
+      json.dump(result_dict, f, indent=4)
 
 
 
@@ -492,6 +519,11 @@ def mainFedAVG(args,cfg):
   # Load Datasets
   datasets = load_datasets(cfg['dataset_dirs'], cfg['image_size'])
   log(INFO, "Datasets loaded. Number of datasets: %s", len(datasets))
+  log(INFO, "Number of Clients: %s", len(datasets))
+
+  for ix in range(len(datasets)):
+    log(INFO,f'Number of samples for client {ix} : {datasets[ix].__len__()}')
+  
   train_dataloaders, val_dataloaders, test_dataloaders = prepare_datasets(
     datasets=datasets, 
     batch_size=cfg['batch_size'], 
@@ -514,9 +546,9 @@ def mainFedAVG(args,cfg):
     
   #Define Strategy
   strategy = fl.server.strategy.FedAvg(
-    fraction_fit=0.0001, 
+    fraction_fit=1.0, 
     min_fit_clients=cfg['num_clients_per_round_fit'], 
-    fraction_evaluate=0.0001, 
+    fraction_evaluate=1.0, 
     min_evaluate_clients=cfg['num_clients_per_round_eval'], 
     min_available_clients=cfg['num_clients'], 
     on_fit_config_fn=get_on_fit_config_function(cfg['config_fit']),
@@ -545,10 +577,12 @@ def mainFedAVG(args,cfg):
   except Exception as e:
     print(f"While simulating an error has occured : {e}")
     traceback.print_exc()
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
     exit()
   finally:
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
   
   results = []
   test_model = UNet(
@@ -571,10 +605,15 @@ def mainFedAVG(args,cfg):
         dice_scores.append(dice)
     results.append(torch.tensor(dice_scores).mean().item())
   
-  with open(os.path.join(cfg['output_dir'],'results.txt'),'w') as f:
-    for idx in range(cfg['num_clients']):
-      if idx > 0 : f.write(f'\nclient {idx} has dice score :{results[idx]}')
-      else : f.write(f'client {idx} has dice score {results[idx]}')
+  result_dict = {
+    '0':-1.0,
+    '1':-1.0,
+    '2':-1.0
+  }
+  for idx in range(cfg['num_clients']):
+      result_dict[str(idx)] = results[idx]
+  with open(os.path.join(cfg['output_dir'],f'results{args.run}.json'), "w") as f:
+      json.dump(result_dict, f, indent=4)
 
 
 
@@ -592,6 +631,11 @@ def mainFedAVGM(args,cfg):
   # Load Datasets
   datasets = load_datasets(cfg['dataset_dirs'], cfg['image_size'])
   log(INFO, "Datasets loaded. Number of datasets: %s", len(datasets))
+  log(INFO, "Number of Clients: %s", len(datasets))
+
+  for ix in range(len(datasets)):
+    log(INFO,f'Number of samples for client {ix} : {datasets[ix].__len__()}')
+  
   train_dataloaders, val_dataloaders, test_dataloaders = prepare_datasets(
     datasets=datasets, 
     batch_size=cfg['batch_size'], 
@@ -614,9 +658,9 @@ def mainFedAVGM(args,cfg):
     
   #Define Strategy
   strategy = fl.server.strategy.FedAvgM(
-    fraction_fit=0.0001, 
+    fraction_fit=1.0, 
     min_fit_clients=cfg['num_clients_per_round_fit'], 
-    fraction_evaluate=0.0001, 
+    fraction_evaluate=1.0, 
     min_evaluate_clients=cfg['num_clients_per_round_eval'], 
     min_available_clients=cfg['num_clients'], 
     on_fit_config_fn=get_on_fit_config_function(cfg['config_fit']),
@@ -645,10 +689,12 @@ def mainFedAVGM(args,cfg):
   except Exception as e:
     print(f"While simulating an error has occured : {e}")
     traceback.print_exc()
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
     exit()
   finally:
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
   
   results = []
   test_model = UNet(
@@ -671,10 +717,15 @@ def mainFedAVGM(args,cfg):
         dice_scores.append(dice)
     results.append(torch.tensor(dice_scores).mean().item())
   
-  with open(os.path.join(cfg['output_dir'],'results.txt'),'w') as f:
-    for idx in range(cfg['num_clients']):
-      if idx > 0 : f.write(f'\nclient {idx} has dice score :{results[idx]}')
-      else : f.write(f'client {idx} has dice score {results[idx]}')
+  result_dict = {
+    '0':-1.0,
+    '1':-1.0,
+    '2':-1.0
+  }
+  for idx in range(cfg['num_clients']):
+      result_dict[str(idx)] = results[idx]
+  with open(os.path.join(cfg['output_dir'],f'results{args.run}.json'), "w") as f:
+      json.dump(result_dict, f, indent=4)
 
 def mainFedADAGRAD(args,cfg):
   # Parse config file and print it out
@@ -690,6 +741,11 @@ def mainFedADAGRAD(args,cfg):
   # Load Datasets
   datasets = load_datasets(cfg['dataset_dirs'], cfg['image_size'])
   log(INFO, "Datasets loaded. Number of datasets: %s", len(datasets))
+  log(INFO, "Number of Clients: %s", len(datasets))
+
+  for ix in range(len(datasets)):
+    log(INFO,f'Number of samples for client {ix} : {datasets[ix].__len__()}')
+  
   train_dataloaders, val_dataloaders, test_dataloaders = prepare_datasets(
     datasets=datasets, 
     batch_size=cfg['batch_size'], 
@@ -715,7 +771,9 @@ def mainFedADAGRAD(args,cfg):
     num_classes=cfg['num_classes'],
     random_seed=cfg['random_seed']
   )
-  initial_parameters = fl.common.ndarrays_to_parameters([val.cpu().numpy() for val in initial_model.state_dict().values()])
+  initial_parameters = fl.common.ndarrays_to_parameters(
+    [val.cpu().numpy() for val in initial_model.state_dict().values()]
+  )
   
   #Define Strategy
   strategy = fl.server.strategy.FedAdagrad(
@@ -753,34 +811,43 @@ def mainFedADAGRAD(args,cfg):
   except Exception as e:
     print(f"While simulating an error has occured : {e}")
     traceback.print_exc()
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
     exit()
   finally:
-    os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
+    if os.path.exists(os.path.join(cfg['output_dir'],'best_dice.json')):
+      os.remove(os.path.join(cfg['output_dir'],'best_dice.json'))
   
   results = []
-  model = UNet(
+  test_model = UNet(
     in_channels=cfg['input_channels'],
     num_classes=cfg['num_classes'],
     random_seed=cfg['random_seed']
   )
+  test_model.load_state_dict(torch.load(os.path.join(cfg['output_dir'],'fedAVG.pth')))
   for idx in range(cfg['num_clients']):
-    model.eval()
-    model.to(device)
-    
+    test_model.eval()
+    test_model.to(device)
+    dice_scores = []
     with torch.no_grad():
       loop = tqdm(test_dataloaders[idx])
       
       for images, masks in loop:
         images, masks = images.to(device), masks.to(device)
-        outputs = model(images)
+        outputs = test_model(images)
         iou, dice = iou_dice_score(outputs, masks)
-        results.append(dice)
+        dice_scores.append(dice)
+    results.append(torch.tensor(dice_scores).mean().item())
   
-  with open(os.path.join(cfg['output_dir'],'results.txt'),'w') as f:
-    for idx in range(cfg['num_clients']):
-      if idx > 0 : f.write(f'\nclient {idx} has dice score :{results[idx]}')
-      else : f.write(f'client {idx} has dice score {results[idx]}')
+  result_dict = {
+    '0':-1.0,
+    '1':-1.0,
+    '2':-1.0
+  }
+  for idx in range(cfg['num_clients']):
+      result_dict[str(idx)] = results[idx]
+  with open(os.path.join(cfg['output_dir'],f'results{args.run}.json'), "w") as f:
+      json.dump(result_dict, f, indent=4)
 
             
 if __name__ == "__main__":
@@ -812,6 +879,8 @@ if __name__ == "__main__":
     mainFedPER(args,cfg)
   elif args.strategy == 'fedAVG':
     mainFedAVG(args,cfg)
+  elif args.strategy == 'fedAVGM':
+    mainFedAVGM(args,cfg)
   elif args.strategy == 'fedADAGRAD':
     mainFedADAGRAD(args,cfg)
   else :
